chore: fix telemetry -> crowdsourced renames
