+++ common/src/main/java/com/wynntils/core/crowdsource/datatype/LootrunTaskLocation.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.core.telemetry.datatype;
+package com.wynntils.core.crowdsource.datatype;
 
 import com.google.common.collect.ComparisonChain;
 import com.wynntils.models.lootrun.type.LootrunLocation;
