+++ common/src/main/java/com/wynntils/core/crowdsource/type/CrowdSourcedDataType.java
@@ -2,9 +2,9 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.core.telemetry.type;
+package com.wynntils.core.crowdsource.type;
 
-import com.wynntils.core.telemetry.datatype.LootrunTaskLocation;
+import com.wynntils.core.crowdsource.datatype.LootrunTaskLocation;
 
 /**
  * This enum represents the type of crowd sourced data that is being collected.
