+++ common/src/main/java/com/wynntils/core/json/JsonManager.java
@@ -13,7 +13,7 @@ import com.google.gson.JsonParser;
 import com.google.gson.stream.JsonReader;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Manager;
-import com.wynntils.core.telemetry.CrowdSourcedData;
+import com.wynntils.core.crowdsource.CrowdSourcedData;
 import com.wynntils.utils.EnumUtils;
 import com.wynntils.utils.FileUtils;
 import com.wynntils.utils.colors.CustomColor;
